# swagger.yaml
openapi: 3.0.0
info:
  title: Airtime2Cash
  description: In order to convert mobile airtime into cash using the web-based application Airtime2Cash, you must first initiate the process from the platform. Once our admin has confirmed the transaction as successful, 70% of the transferred amount will be credited to the client's wallet, from which the client can withdraw the money to his or her bank account.
  contact:
    name: SQ11B Pod F
    email: airtime2cashpodf@gmail.com
  version: '1.0'
servers:
- url: http://localhost:7000
  description: This is the localhost development server
  variables: {}
paths:
...
  /user/singleUser/{id}:
    parameters:
      - in: path
        name: id
        schema:
          type: string
        required: true
        description: The user ID
    get:
      tags:
      - User
      summary: Single User
      description: This returns details of a single user
      operationId: SingleUser
      parameters: []
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
      security: []
...
  /user/resetPassword:
    patch:
      tags:
      - Auth
      summary: Forgot Password Reset
      description: After the user's email has been verified to be valid, the user can then reset his or her password
      operationId: ForgotPasswordReset
      parameters: []
      requestBody:
        description: ''
        content:
          application/json:
            schema:
              allOf:
              - $ref: '#/components/schemas/restepasswordrequest'
              - example:
                  token: d55e5e01-7b41-440f-961c-a16ab779e00a
                  password: mypassword
                  confirmPassword: mypassword
            example:
              token: d55e5e01-7b41-440f-961c-a16ab779e00a
              password: mypassword
              confirmPassword: mypassword
        required: true
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
      security: []
  /user/register:
    post:
      tags:
      - Auth
      summary: Register User
      description: This is the signup endpoint, which allows you to register a user
      operationId: RegisterUser
      parameters: []
      requestBody:
        description: ''
        content:
          application/json:
            schema:
              allOf:
              - $ref: '#/components/schemas/registeruserrequest'
              - example:
                  firstName: Adeyinka
                  lastName: Kayode
                  userName: tjvpro
                  email: example@domain.com
                  phoneNumber: '08093059174'
                  password: '1234567890'
                  confirmPassword: '1234567890'
            example:
              firstName: Adeyinka
              lastName: Kayode
              userName: tjvpro
              email: example@domain.com
              phoneNumber: '08093059174'
              password: '1234567890'
              confirmPassword: '1234567890'
        required: true
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
      security: []
  /user/login:
    post:
      tags:
      - Auth
      summary: Login User
      description: You can log in either by using a registered email or a registered username with a password
      operationId: LoginUser
      parameters: []
      requestBody:
        description: ''
        content:
          application/json:
            schema:
              allOf:
              - $ref: '#/components/schemas/loginuserrequest'
              - example:
                  emailOrUsername: email@hostname.com
                  password: userpassword
            example:
              emailOrUsername: email@hostname.com
              password: userpassword
        required: true
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
      security: []
  /user/verify/{token}:
    parameters:
      - in: path
        name: token
        schema:
          type: string
        required: true
        description: The user ID
    get:
      tags:
      - Auth
      summary: Verify Email
      description: This endpoint verifies an unverified email, to get the token for verification you can go through the resend verification token endpoint or upon registration of the user, after which the user would be redirected to the login page
      operationId: VerifyEmail
      parameters: []
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
      security: []
  /user/forgetPassword:
    post:
      tags:
      - Auth
      summary: Forgot Password Token
      description: This endpoint generates a token that is sent to the email provided. Then the token can be used to reset the password via the reset password endpoint
      operationId: ForgotPasswordToken
      parameters: []
      requestBody:
        description: ''
        content:
          application/json:
            schema:
              allOf:
              - $ref: '#/components/schemas/forgetpasswordrequest'
              - example:
                  email: example@domain.com
            example:
              email: example@domain.com
        required: true
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
...
  /user/deleteUser/{userId}:
    parameters:
      - in: path
        name: userId
        schema:
          type: string
        required: true
        description: The user ID
    delete:
      tags:
      - Admin
      summary: Delete User
      description: This endpoint deletes a registered user from the system database, accessed only by logged-in admin
      operationId: DeleteUser
      parameters: []
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
      security:
      - bearer: []
components:
  schemas:
...
    restepasswordrequest:
      title: restepasswordrequest
      required:
      - token
      - password
      - confirmPassword
      type: object
      properties:
        token:
          type: string
        password:
          type: string
        confirmPassword:
          type: string
      example:
        token: d55e5e01-7b41-440f-961c-a16ab779e00a
        password: mypassword
        confirmPassword: mypassword
    registeruserrequest:
      title: registeruserrequest
      required:
      - firstName
      - lastName
      - userName
      - email
      - phoneNumber
      - password
      - confirmPassword
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        userName:
          type: string
        email:
          type: string
        phoneNumber:
          type: string
        password:
          type: string
        confirmPassword:
          type: string
      example:
        firstName: Adeyinka
        lastName: Kayode
        userName: tjvpro
        email: example@domain.com
        phoneNumber: '08093059174'
        password: '1234567890'
        confirmPassword: '1234567890'
    loginuserrequest:
      title: loginuserrequest
      required:
      - emailOrUsername
      - password
      type: object
      properties:
        emailOrUsername:
          type: string
        password:
          type: string
      example:
        emailOrUsername: example@domain.com
        password: '1234567890'
    forgetpasswordrequest:
      title: forgetpasswordrequest
      required:
      - email
      type: object
      properties:
        email:
          type: string
      example:
        email: example@domain.com
...
  securitySchemes:
    bearer:
      type: http
      scheme: bearer
security: []
...